% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputHelpers.R
\name{combine_chains}
\alias{combine_chains}
\title{Combine Chains
Function to combine the posterior samples from the multiple chains.}
\usage{
combine_chains(results)
}
\arguments{
\item{results}{A list of MCMC chain results.}
}
\value{
A list with combined results, including median, shift, first quartile, and asymptote values.
}
\description{
Combine Chains
Function to combine the posterior samples from the multiple chains.
}
\examples{
combine_results <- combine_chains(mcmc_results)

}
